Bug fixes and improvements
