--- app/backend/core/message_truncater/function_format.py
@@ -1,64 +1,0 @@
-# Based on https://github.com/forestwanglin/openai-java/blob/main/jtokkit/src/main/java/xyz/felh/openai/jtokkit/utils/TikTokenUtils.java
-
-
-def format_function_definitions(tools):
-    lines = []
-    lines.append("namespace functions {")
-    lines.append("")
-    for tool in tools:
-        function = tool.get("function")
-        if function_description := function.get("description"):
-            lines.append(f"// {function_description}")
-        function_name = function.get("name")
-        parameters = function.get("parameters", {})
-        properties = parameters.get("properties")
-        if properties and properties.keys():
-            lines.append(f"type {function_name} = (_: {{")
-            lines.append(format_object_parameters(parameters, 0))
-            lines.append("}) => any;")
-        else:
-            lines.append(f"type {function_name} = () => any;")
-        lines.append("")
-    lines.append("} // namespace functions")
-    return "\n".join(lines)
-
-
-def format_object_parameters(parameters, indent):
-    properties = parameters.get("properties")
-    if not properties:
-        return ""
-    required_params = parameters.get("required", [])
-    lines = []
-    for key, props in properties.items():
-        description = props.get("description")
-        if description:
-            lines.append(f"// {description}")
-        question = "?"
-        if required_params and key in required_params:
-            question = ""
-        lines.append(f"{key}{question}: {format_type(props, indent)},")
-    return "\n".join([" " * max(0, indent) + line for line in lines])
-
-
-def format_type(props, indent):
-    type = props.get("type")
-    if type == "string":
-        if "enum" in props:
-            return " | ".join([f'"{item}"' for item in props["enum"]])
-        return "string"
-    elif type == "array":
-        # items is required, OpenAI throws an error if it's missing
-        return f"{format_type(props['items'], indent)}[]"
-    elif type == "object":
-        return f"{{\n{format_object_parameters(props, indent + 2)}\n}}"
-    elif type in ["integer", "number"]:
-        if "enum" in props:
-            return " | ".join([f'"{item}"' for item in props["enum"]])
-        return "number"
-    elif type == "boolean":
-        return "boolean"
-    elif type == "null":
-        return "null"
-    else:
-        # This is a guess, as an empty string doesn't yield the expected token count
-        return "any"
